requirements.txt
tensorflow
dvc
pandas
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
mlflow==1.26.1
gdown
-e .
